+++ public/sw.js
@@ -6,26 +6,33 @@ self.addEventListener('push', function (event) {
 
     if (event.data) {
         const msg = event.data.json();
-        console.log(msg)
         event.waitUntil(
             self.registration.showNotification(msg.title, {
                 body: msg.body,
                 icon: msg.icon,
-                actions: msg.actions
+                actions: msg.actions,
+                data: msg.data
             })
         );
     }
 });
 
-self.addEventListener('notificationclick', function(event) {
-    console.log(event.notification.data);
+self.addEventListener('notificationclick', function (event) {
 
     event.notification.close();
 
-    if (event.action === 'events') {
-        clients.openWindow("/events");
-    } else if (event.action === 'news') {
-        clients.openWindow("/news");
+    if (event.action === 'event-detail') {
+        if (event.notification.data.id) {
+            clients.openWindow("/events/" + event.notification.data.id + "/detail");
+        } else {
+            clients.openWindow("/events");
+        }
+    } else if (event.action === 'news-detail') {
+        if (event.notification.data.id) {
+            clients.openWindow("/news/" + event.notification.data.id + "/detail");
+        } else {
+            clients.openWindow("/news");
+        }
     } else {
         clients.openWindow("/");
     }